feat: honor required params
