+++ libraries/gherkin-http-api-generator/src/main/java/ht/eyfout/junit/jupiter/gherkin/http/SwaggerAPI.java
@@ -20,10 +20,8 @@ record SwaggerAPI(String path, PathItem.HttpMethod method, Operation operation) 
     }
 
     public String description() {
-        return Optional.ofNullable(operation.getDescription()).orElseGet(() ->
-                Optional.ofNullable(operation.getSummary()).orElseGet(() ->
-                        operation.toString()
-                )
+        return Optional.ofNullable(operation.getSummary()).orElseGet(() ->
+                Optional.ofNullable(operation.getDescription()).orElse(null)
         );
     }
