+++ settings.gradle
@@ -5,4 +5,4 @@ include(":libraries:gherkin-http-provider")
 include(":libraries:gherkin-http-api-generator")
 
 include(":examples:micronaut-app")
-include(":plugins")
+include(":plugins:junit-gherkin-http")
